+++ tests/SearchEndpoint/FilterEndpointTest.php
@@ -11,8 +11,8 @@
 
 namespace ONGR\ElasticsearchDSL\Tests\Unit\SearchEndpoint;
 
-use ONGR\ElasticsearchDSL\Filter\MatchAllFilter;
 use ONGR\ElasticsearchDSL\Query\FilteredQuery;
+use ONGR\ElasticsearchDSL\Query\MatchAllQuery;
 use ONGR\ElasticsearchDSL\SearchEndpoint\FilterEndpoint;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use PHPUnit_Framework_MockObject_MockObject as MockObject;
@@ -55,7 +55,7 @@ class FilterEndpointTest extends \PHPUnit_Framework_TestCase
         $this->assertNull($instance->normalize($normalizerInterface));
         $this->assertFalse($instance->hasReference('filtered_query'));
 
-        $matchAllFilter = new MatchAllFilter();
+        $matchAllFilter = new MatchAllQuery();
         $instance->add($matchAllFilter);
 
         $this->assertNull($instance->normalize($normalizerInterface));
@@ -81,7 +81,7 @@ class FilterEndpointTest extends \PHPUnit_Framework_TestCase
     public function testEndpointGetter()
     {
         $filterName = 'acme_filter';
-        $filter = new MatchAllFilter();
+        $filter = new MatchAllQuery();
         $endpoint = new FilterEndpoint();
         $endpoint->add($filter, $filterName);
         $builders = $endpoint->getAll();